Fix signature and palette spread source
