+++ Common/src/main/java/dev/lukebemish/dynamicassetgenerator/api/client/generators/texsources/AnimationFrameCapture.java
@@ -33,7 +33,7 @@ public final class AnimationFrameCapture implements TexSource {
     }
 
     @Override
-    public Codec<? extends TexSource> codec() {
+    public @NotNull Codec<? extends TexSource> codec() {
         return CODEC;
     }
