+++ Common/src/main/java/dev/lukebemish/dynamicassetgenerator/api/client/generators/texsources/AnimationSplittingSource.java
@@ -17,6 +17,7 @@ import dev.lukebemish.dynamicassetgenerator.impl.util.Maath;
 import dev.lukebemish.dynamicassetgenerator.impl.util.MultiCloser;
 import net.minecraft.server.packs.resources.IoSupplier;
 import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.Closeable;
@@ -43,7 +44,7 @@ public final class AnimationSplittingSource implements TexSource {
     }
 
     @Override
-    public Codec<? extends TexSource> codec() {
+    public @NotNull Codec<? extends TexSource> codec() {
         return CODEC;
     }
