+++ Common/src/main/java/dev/lukebemish/dynamicassetgenerator/api/client/generators/texsources/ChannelRouteSource.java
@@ -17,6 +17,7 @@ import dev.lukebemish.dynamicassetgenerator.api.colors.ColorTypes;
 import dev.lukebemish.dynamicassetgenerator.api.colors.operations.PointwiseOperation;
 import dev.lukebemish.dynamicassetgenerator.impl.util.MultiCloser;
 import net.minecraft.server.packs.resources.IoSupplier;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
@@ -108,7 +109,7 @@ public class ChannelRouteSource implements TexSource {
     }
 
     @Override
-    public Codec<? extends TexSource> codec() {
+    public @NotNull Codec<? extends TexSource> codec() {
         return CODEC;
     }
