+++ Common/src/main/java/dev/lukebemish/dynamicassetgenerator/api/client/generators/texsources/CropSource.java
@@ -14,6 +14,7 @@ import dev.lukebemish.dynamicassetgenerator.api.client.generators.TexSourceDataH
 import dev.lukebemish.dynamicassetgenerator.api.client.image.ImageUtils;
 import dev.lukebemish.dynamicassetgenerator.impl.client.NativeImageHelper;
 import net.minecraft.server.packs.resources.IoSupplier;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
@@ -47,7 +48,7 @@ public final class CropSource implements TexSource {
         this.input = input;
     }
 
-    public Codec<CropSource> codec() {
+    public @NotNull Codec<? extends TexSource> codec() {
         return CODEC;
     }
