+++ Common/src/main/java/dev/lukebemish/dynamicassetgenerator/api/client/generators/texsources/FallbackSource.java
@@ -12,6 +12,7 @@ import dev.lukebemish.dynamicassetgenerator.api.ResourceGenerationContext;
 import dev.lukebemish.dynamicassetgenerator.api.client.generators.TexSource;
 import dev.lukebemish.dynamicassetgenerator.api.client.generators.TexSourceDataHolder;
 import net.minecraft.server.packs.resources.IoSupplier;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.helpers.NOPLogger;
@@ -35,7 +36,7 @@ public final class FallbackSource implements TexSource {
         this.fallback = fallback;
     }
 
-    public Codec<FallbackSource> codec() {
+    public @NotNull Codec<? extends TexSource> codec() {
         return CODEC;
     }
