+++ Common/src/main/java/dev/lukebemish/dynamicassetgenerator/api/client/generators/texsources/OverlaySource.java
@@ -9,6 +9,7 @@ import com.mojang.serialization.Codec;
 import dev.lukebemish.dynamicassetgenerator.api.client.generators.TexSource;
 import dev.lukebemish.dynamicassetgenerator.api.colors.operations.ColorOperations;
 import dev.lukebemish.dynamicassetgenerator.api.colors.operations.PointwiseOperation;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.Objects;
@@ -24,7 +25,7 @@ public class OverlaySource extends AbstractManyOperationSource {
     }
 
     @Override
-    public Codec<? extends TexSource> codec() {
+    public @NotNull Codec<? extends TexSource> codec() {
         return CODEC;
     }
