+++ Common/src/main/java/dev/lukebemish/dynamicassetgenerator/api/client/generators/texsources/mask/AddMask.java
@@ -10,6 +10,7 @@ import dev.lukebemish.dynamicassetgenerator.api.client.generators.TexSource;
 import dev.lukebemish.dynamicassetgenerator.api.client.generators.texsources.AbstractManyOperationSource;
 import dev.lukebemish.dynamicassetgenerator.api.colors.operations.ColorOperations;
 import dev.lukebemish.dynamicassetgenerator.api.colors.operations.PointwiseOperation;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.Objects;
@@ -26,7 +27,7 @@ public class AddMask extends AbstractManyOperationSource {
     }
 
     @Override
-    public Codec<? extends TexSource> codec() {
+    public @NotNull Codec<? extends TexSource> codec() {
         return CODEC;
     }
