+++ Common/src/main/java/dev/lukebemish/dynamicassetgenerator/api/client/generators/texsources/mask/EdgeMask.java
@@ -50,7 +50,7 @@ public final class EdgeMask implements TexSource {
     }
 
     @Override
-    public Codec<? extends TexSource> codec() {
+    public @NotNull Codec<? extends TexSource> codec() {
         return CODEC;
     }
